[TAN-1480] Disable submit button in modal to showcase how feature works
