+++ front/app/containers/Admin/projects/project/offlineInputs/TopBar/ImportButtons.tsx
@@ -6,12 +6,8 @@ import {
   Button,
   Dropdown,
   DropdownListItem,
-  TooltipContentWrapper,
   colors,
 } from '@citizenlab/cl2-component-library';
-import Tippy from '@tippyjs/react';
-
-import useFeatureFlag from 'hooks/useFeatureFlag';
 
 import { FormattedMessage } from 'utils/cl-intl';
 
@@ -24,9 +20,6 @@ interface Props {
 
 const ImportButtons = ({ onClickPDFImport, onClickExcelImport }: Props) => {
   const [dropdownOpened, setDropdownOpened] = useState(false);
-  const importPrintedFormsEnabled = useFeatureFlag({
-    name: 'import_printed_forms',
-  });
 
   return (
     <Box>
@@ -48,35 +41,15 @@ const ImportButtons = ({ onClickPDFImport, onClickExcelImport }: Props) => {
             <DropdownListItem onClick={onClickExcelImport}>
               <FormattedMessage {...messages.importExcelFile} />
             </DropdownListItem>
-            <DropdownListItem
-              onClick={importPrintedFormsEnabled ? onClickPDFImport : undefined}
-            >
-              <Tippy
-                interactive={true}
-                theme={''}
-                maxWidth={350}
-                content={
-                  <TooltipContentWrapper tippytheme="light">
-                    <FormattedMessage {...messages.disabledPDFImportTooltip} />
-                  </TooltipContentWrapper>
-                }
-              >
-                <Box
-                  display="flex"
-                  style={
-                    importPrintedFormsEnabled ? {} : { cursor: 'not-allowed' }
-                  }
-                >
-                  <Box mr="12px" alignSelf="center">
-                    <FormattedMessage {...messages.importPDFFile} />
-                  </Box>
-                  <Box alignSelf="center">
-                    <Badge color={colors.teal400}>
-                      <FormattedMessage {...messages.premium} />
-                    </Badge>
-                  </Box>
-                </Box>
-              </Tippy>
+            <DropdownListItem onClick={onClickPDFImport}>
+              <Box mr="12px" alignSelf="center">
+                <FormattedMessage {...messages.importPDFFile} />
+              </Box>
+              <Box alignSelf="center">
+                <Badge color={colors.teal400}>
+                  <FormattedMessage {...messages.premium} />
+                </Badge>
+              </Box>
             </DropdownListItem>
           </>
         }
